2/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.29
django-bootstrap4==2.2.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==3.0.2
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
